pyproject.toml
[project]
name = "ecommecer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "dependencies>=7.7.1",
    "django>=5.2.5",
    "django-allauth>=65.10.0",
    "django-ninja>=1.4.3",
    "django-ninja-jwt>=5.3.7",
    "djangorestframework-simplejwt>=5.5.1",
    "fastapi>=0.116.1",
    "jwt>=1.4.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "zxcvbn>=4.5.0",
]
